Change directory hierarchy according to DDD
